Add tests for Proto DataStore and Preference DataStore
